+++ CHANGELOG.md
@@ -3,8 +3,12 @@
 - Improved commands logging.
 - Fix passing of release bundle to finalize commands.
 - Handler `URL` request errors.
-- `ReleasePackerCommandUploadReleaseBundle``: fix parameter `file` format.
-- Server: Fix parse of request credential.
+- `ReleasePackerCommandUploadReleaseBundle``:
+  - Fix parameter `file` format.
+  - Fix parameter `release`.
+- Server:
+  - Fix parse of request credential.
+  - Append uploaded releases to `releases-file`.
 
 ## 1.0.11